Django_2/SQL/d05/requirements.txt
asgiref==3.8.1
csvfile==3.4.0
Django==5.1.2
Markdown==3.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
